tweak: Moved package search test to its own file
