--- tests/stackmachine/stackmachine-sdk.test.ts
@@ -1,558 +1,0 @@
-import {
-  createZip,
-  type AppAlias,
-  type DeployApp,
-  type DeploymentCreateInput,
-  type Log,
-  type SearchPackageVersion,
-} from "stackmachine";
-import {
-  AppInfo,
-  buildTempDir,
-  randomAppName,
-  sleep,
-  TestEnv,
-} from "../../src";
-import { generateNeedlesslySecureRandomPassword } from "../../src/security";
-
-jest.setTimeout(600_000);
-
-type StackMachineClient = Awaited<ReturnType<TestEnv["stackmachineSdk"]>>;
-
-function appFetchTarget(app: Pick<DeployApp, "id" | "url">): AppInfo {
-  return {
-    id: app.id,
-    url: app.url,
-    dir: "",
-    app: {
-      id: app.id,
-      url: app.url,
-      permalink: "",
-      activeVersionId: null,
-    },
-    version: {
-      appId: app.id,
-      appVersionId: "",
-      name: "",
-      path: "",
-      url: app.url,
-    },
-  };
-}
-
-async function uploadInlineFiles(
-  client: StackMachineClient,
-  files: Record<string, string>,
-): Promise<string> {
-  const zip = await createZip(files);
-  return client.files.upload(zip);
-}
-
-async function deployUploadedApp(
-  client: StackMachineClient,
-  env: TestEnv,
-  uploadUrl: string,
-  appName = randomAppName(),
-  extraInput: Partial<DeploymentCreateInput> = {},
-): Promise<DeployApp> {
-  const appVersion = await env.deployStackMachineApp(client, {
-    appName,
-    owner: env.namespace,
-    uploadUrl,
-    ...extraInput,
-  });
-  const app = appVersion.app;
-  await env.recordDeployedApp({
-    appId: app.id,
-    appName: app.name,
-    appUrl: app.url,
-    appPermalink: app.url,
-  });
-  return app;
-}
-
-async function deployInlinePhpApp(
-  client: StackMachineClient,
-  env: TestEnv,
-  body: string,
-  appName = randomAppName(),
-  extraInput: Partial<DeploymentCreateInput> = {},
-): Promise<DeployApp> {
-  const uploadUrl = await uploadInlineFiles(client, {
-    "index.php": body,
-  });
-  return deployUploadedApp(client, env, uploadUrl, appName, extraInput);
-}
-
-async function expectAppBody(
-  env: TestEnv,
-  app: DeployApp,
-  expectedSubstring: string,
-  urlOrPath = "/",
-): Promise<void> {
-  const response = await env.fetchApp(appFetchTarget(app), urlOrPath, {
-    forceWait: true,
-  });
-  const body = await response.text();
-  expect(body).toContain(expectedSubstring);
-}
-
-async function listAppDomains(
-  client: StackMachineClient,
-  appId: string,
-): Promise<AppAlias[]> {
-  return (await client.apps.domains.list({ app: appId })).data;
-}
-
-async function waitForLogs(
-  client: StackMachineClient,
-  app: DeployApp,
-  substring: string,
-  timeoutMs = 200_000,
-): Promise<void> {
-  const start = Date.now();
-  let latestLogs: Log[] | undefined;
-  while (Date.now() - start < timeoutMs) {
-    if (app.activeVersion) {
-      latestLogs = (
-        await client.apps.versions.logs.list({
-          version: app.activeVersion.id,
-          since: new Date(Date.now() - 30 * 60 * 1000),
-        })
-      ).data;
-    }
-    if (latestLogs?.some((entry) => entry.message.includes(substring))) {
-      return;
-    }
-    await sleep(3_000);
-  }
-  throw new Error(
-    [
-      `Timed out waiting for logs containing '${substring}' for app '${app.id}' (${app.url})`,
-      "Latest logs:",
-      ...(latestLogs ?? []).map((entry) => {
-        const datetime =
-          entry.datetime instanceof Date
-            ? entry.datetime.toISOString()
-            : String(entry.datetime);
-        return `${datetime} ${entry.stream} ${entry.instanceId} ${entry.message}`;
-      }),
-    ].join("\n"),
-  );
-}
-
-async function waitForDeletion(
-  client: StackMachineClient,
-  appId: string,
-  timeoutMs = 60_000,
-): Promise<void> {
-  const start = Date.now();
-  while (Date.now() - start < timeoutMs) {
-    const [app] = await client.apps.retrieveMany([appId]);
-    if (!app) {
-      return;
-    }
-    await sleep(2_000);
-  }
-  throw new Error(`Timed out waiting for deletion of app '${appId}'`);
-}
-
-async function waitForDomainDeletion(
-  client: StackMachineClient,
-  appId: string,
-  aliasId: string,
-  timeoutMs = 60_000,
-): Promise<void> {
-  const start = Date.now();
-  while (Date.now() - start < timeoutMs) {
-    const domains = await listAppDomains(client, appId);
-    if (!domains.find((item) => item.id === aliasId)) {
-      return;
-    }
-    await sleep(2_000);
-  }
-  throw new Error(`Timed out waiting for deletion of domain '${aliasId}'`);
-}
-
-async function addDomain(
-  client: StackMachineClient,
-  env: TestEnv,
-  app: DeployApp,
-): Promise<{ alias: AppAlias; domainName: string }> {
-  const zone = `${crypto.randomUUID().replace(/-/g, "")}.com`;
-  const domainName = `www.${zone}`;
-
-  await env.runWasmerCommand({
-    args: ["domain", "register", zone],
-  });
-
-  const alias = await client.apps.domains.create({
-    app: app.id,
-    hostname: domainName,
-  });
-  expect(alias.expectedDnsRecords.length).toBeGreaterThan(0);
-  return { alias, domainName };
-}
-
-async function waitForDomainPresent(
-  client: StackMachineClient,
-  appId: string,
-  aliasId: string,
-  timeoutMs = 60_000,
-): Promise<AppAlias> {
-  const start = Date.now();
-  while (Date.now() - start < timeoutMs) {
-    const domains = await listAppDomains(client, appId);
-    const alias = domains.find((item) => item.id === aliasId);
-    if (alias) {
-      return alias;
-    }
-    await sleep(2_000);
-  }
-  throw new Error(`Timed out waiting for domain '${aliasId}' to be attached`);
-}
-
-// Publish a fresh, uniquely-named package to the test namespace so search has
-// something deterministic to find. Returns the bare package name (without the
-// owner).
-async function publishUniquePackage(env: TestEnv): Promise<string> {
-  const name = `sdk-search-${crypto.randomUUID().replace(/-/g, "").slice(0, 16)}`;
-  const fullName = `${env.namespace}/${name}`;
-  const pkgDir = await buildTempDir({
-    "wasmer.toml": `
-[package]
-name = "${fullName}"
-version = "0.1.0"
-
-[dependencies]
-"wasmer/static-web-server" = "1"
-
-[fs]
-"/public" = "public"
-
-[[command]]
-name = "script"
-module = "wasmer/static-web-server:webserver"
-runner = "https://webc.org/runner/wasi"
-`,
-    public: {
-      "index.html": name,
-    },
-  });
-  await env.runWasmerCommand({ args: ["publish"], cwd: pkgDir });
-  return name;
-}
-
-// Search indexing is asynchronous after publish, so poll until the package
-// shows up (or time out with the most recent results for debugging).
-async function waitForPackageInSearch(
-  client: StackMachineClient,
-  env: TestEnv,
-  packageName: string,
-  timeoutMs = 120_000,
-): Promise<SearchPackageVersion> {
-  const start = Date.now();
-  let lastSeen: string[] = [];
-  while (Date.now() - start < timeoutMs) {
-    const page = await client.packages.search({
-      query: packageName,
-      filter: { owner: env.namespace },
-    });
-    const match = page.data.find(
-      (result) =>
-        result.package.namespace === env.namespace &&
-        result.package.packageName === packageName,
-    );
-    if (match) {
-      return match;
-    }
-    lastSeen = page.data.map(
-      (result) => `${result.package.namespace}/${result.package.packageName}`,
-    );
-    await sleep(5_000);
-  }
-  throw new Error(
-    [
-      `Timed out waiting for package '${env.namespace}/${packageName}' to appear in search.`,
-      `Last results: ${lastSeen.join(", ") || "(none)"}`,
-    ].join("\n"),
-  );
-}
-
-describe("stackmachine sdk", () => {
-  const cleanupAppIds: string[] = [];
-
-  afterEach(async () => {
-    const env = TestEnv.fromEnv();
-    if (env.shouldPreserveAppsForCurrentTest()) {
-      cleanupAppIds.length = 0;
-      return;
-    }
-
-    while (cleanupAppIds.length > 0) {
-      const appId = cleanupAppIds.pop();
-      if (!appId) {
-        continue;
-      }
-      try {
-        await env.backend.deleteApp(appId);
-      } catch {
-        // Ignore cleanup races when the test already deleted the app.
-      }
-    }
-  });
-
-  test("deployFromFiles example", async () => {
-    const env = TestEnv.fromEnv();
-    const client = await env.stackmachineSdk();
-    const app = await deployInlinePhpApp(
-      client,
-      env,
-      "<html><body><h1>Hello World!</h1></body></html>",
-    );
-    cleanupAppIds.push(app.id);
-
-    expect(app.name).toBeTruthy();
-    await expectAppBody(env, app, "Hello World!");
-  });
-
-  test("deployFromZip example", async () => {
-    const env = TestEnv.fromEnv();
-    const client = await env.stackmachineSdk();
-    const zip = await createZip({
-      "index.php": "<html><body><h1>Hello Zip!</h1></body></html>",
-    });
-    const uploadUrl = await client.files.upload(zip);
-    const app = await deployUploadedApp(client, env, uploadUrl);
-    cleanupAppIds.push(app.id);
-
-    await expectAppBody(env, app, "Hello Zip!");
-  });
-
-  test("deployFromZipUrl example", async () => {
-    const env = TestEnv.fromEnv();
-    const client = await env.stackmachineSdk();
-    const uploadUrl = await uploadInlineFiles(client, {
-      "index.php": "<html><body><h1>Hello Zip URL!</h1></body></html>",
-    });
-    const app = await deployUploadedApp(client, env, uploadUrl);
-    cleanupAppIds.push(app.id);
-
-    await expectAppBody(env, app, "Hello Zip URL!");
-  });
-
-  test("redeployFromFiles example", async () => {
-    const env = TestEnv.fromEnv();
-    const client = await env.stackmachineSdk();
-    const appName = randomAppName();
-
-    const app1 = await deployInlinePhpApp(
-      client,
-      env,
-      "<html><body><h1>Hello World v1!</h1></body></html>",
-      appName,
-    );
-    cleanupAppIds.push(app1.id);
-    await expectAppBody(env, app1, "Hello World v1!");
-
-    const app2 = await deployInlinePhpApp(
-      client,
-      env,
-      "<html><body><h1>Hello World v2!</h1></body></html>",
-      appName,
-      { allowExistingApp: true },
-    );
-
-    expect(app2.id).toBe(app1.id);
-    await expectAppBody(env, app2, "Hello World v2!");
-  });
-
-  test("deployPerishable example", async () => {
-    const env = TestEnv.fromEnv();
-    const client = await env.stackmachineSdk();
-    const app = await deployInlinePhpApp(
-      client,
-      env,
-      "<html><body><h1>This app will perish in 2 hours</h1></body></html>",
-      randomAppName(),
-      { perishAt: "PT2H" },
-    );
-    cleanupAppIds.push(app.id);
-
-    expect(app.willPerishAt).not.toBeNull();
-    expect(app.willPerishAt!.getTime()).toBeGreaterThan(Date.now());
-    await expectAppBody(env, app, "This app will perish in 2 hours");
-  });
-
-  test("getAppById example", async () => {
-    const env = TestEnv.fromEnv();
-    const client = await env.stackmachineSdk();
-    const app = await deployInlinePhpApp(
-      client,
-      env,
-      "<html><body><h1>Get By Id</h1></body></html>",
-    );
-    cleanupAppIds.push(app.id);
-
-    const fetched = await client.apps.retrieve(app.id);
-    expect(fetched.id).toBe(app.id);
-    expect(fetched.url).toBe(app.url);
-  });
-
-  test("getAppByName example", async () => {
-    const env = TestEnv.fromEnv();
-    const client = await env.stackmachineSdk();
-    const app = await deployInlinePhpApp(
-      client,
-      env,
-      "<html><body><h1>Get By Name</h1></body></html>",
-    );
-    cleanupAppIds.push(app.id);
-
-    const fetched = await client.apps.retrieveByName(app.name, env.namespace);
-    expect(fetched.id).toBe(app.id);
-  });
-
-  test("getAppLogs example", async () => {
-    const env = TestEnv.fromEnv();
-    const client = await env.stackmachineSdk();
-    const app = await deployInlinePhpApp(
-      client,
-      env,
-      `<?php error_log("hello from stackmachine sdk logs"); echo "ok"; ?>`,
-    );
-    cleanupAppIds.push(app.id);
-
-    await expectAppBody(env, app, "ok");
-    await waitForLogs(client, app, "hello from stackmachine sdk logs");
-  });
-
-  test("addAppDomain example", async () => {
-    const env = TestEnv.fromEnv();
-    const client = await env.stackmachineSdk();
-    const app = await deployInlinePhpApp(
-      client,
-      env,
-      "<html><body><h1>Domain Ready</h1></body></html>",
-    );
-    cleanupAppIds.push(app.id);
-
-    const { alias, domainName } = await addDomain(client, env, app);
-    const refreshed = await waitForDomainPresent(client, app.id, alias.id);
-    expect(refreshed.url).toContain(domainName);
-  });
-
-  test("deleteAppDomain example", async () => {
-    const env = TestEnv.fromEnv();
-    const client = await env.stackmachineSdk();
-    const app = await deployInlinePhpApp(
-      client,
-      env,
-      "<html><body><h1>Delete Domain</h1></body></html>",
-    );
-    cleanupAppIds.push(app.id);
-
-    const { alias } = await addDomain(client, env, app);
-    await waitForDomainPresent(client, app.id, alias.id);
-    await client.apps.domains.del(alias.id);
-    await waitForDomainDeletion(client, app.id, alias.id);
-  });
-
-  test("deleteApp example", async () => {
-    const env = TestEnv.fromEnv();
-    const client = await env.stackmachineSdk();
-    const app = await deployInlinePhpApp(
-      client,
-      env,
-      "<html><body><h1>Delete Me</h1></body></html>",
-    );
-
-    await client.apps.del(app.id);
-    await waitForDeletion(client, app.id);
-  });
-
-  test("tiinyBlocked example", async () => {
-    const env = TestEnv.fromEnv();
-    const client = await env.stackmachineSdk();
-    const app = await deployInlinePhpApp(
-      client,
-      env,
-      "<html><body><h1>Hello Tiiny!</h1></body></html>",
-    );
-    cleanupAppIds.push(app.id);
-
-    await expectAppBody(env, app, "Hello Tiiny!");
-  });
-
-  test("deployWordpress example", async () => {
-    const env = TestEnv.fromEnv();
-    const client = await env.stackmachineSdk();
-    const appName = randomAppName();
-
-    const appVersion = await env.deployStackMachineApp(client, {
-      appName,
-      owner: env.namespace,
-      repoUrl: "https://github.com/wordpress/wordpress",
-      branch: "6.8.3",
-      enableDatabase: true,
-      extraData: {
-        wordpress: {
-          adminEmail: "admin@example.com",
-          adminPassword: generateNeedlesslySecureRandomPassword(),
-          adminUsername: "admin",
-          language: "en_US",
-          siteName: "Gallant Goldberg",
-        },
-      },
-    });
-    const app = appVersion.app;
-    await env.recordDeployedApp({
-      appId: app.id,
-      appName,
-      appUrl: app.url,
-      appPermalink: app.url,
-    });
-    cleanupAppIds.push(app.id);
-
-    expect(app.adminUrl).toBeTruthy();
-    const response = await env.fetchApp(appFetchTarget(app), "/", {
-      forceWait: true,
-      noAssertSuccess: true,
-    });
-    expect([200, 302]).toContain(response.status);
-  });
-
-  test("searchPackagesByOwner example", async () => {
-    const env = TestEnv.fromEnv();
-    const client = await env.stackmachineSdk();
-
-    const name = await publishUniquePackage(env);
-
-    const result = await waitForPackageInSearch(client, env, name);
-    expect(result.package.namespace).toBe(env.namespace);
-    expect(result.package.packageName).toBe(name);
-    expect(result.version).toBe("0.1.0");
-    expect(result.createdAt).toBeInstanceOf(Date);
-    expect(result.package.lastVersion?.version).toBe("0.1.0");
-
-    const ownerScoped = await client.packages.search({
-      filter: { owner: env.namespace },
-      limit: 50,
-    });
-    // a package for that owner must exist
-    expect(ownerScoped.data.length).toBeGreaterThan(0);
-    // and all returned packages must belong to the queried owner
-    for (const entry of ownerScoped.data) {
-      expect(entry.package.namespace).toBe(env.namespace);
-    }
-
-    const otherOwner = await client.packages.search({
-      query: name,
-      filter: { owner: "wasmer" },
-    });
-    // the package should *not* be returned when filtering for a different owner
-    expect(
-      otherOwner.data.some((entry) => entry.package.packageName === name),
-    ).toBe(false);
-  });
-});
